Fix the type of _lf_zero_delay_cycle_action_table_size iterator to be size_t + automatic formatter

+++ core/federated/federate.c
@@ -296,7 +296,15 @@ static void update_last_known_status_on_input_port(environment_t* env, tag_t tag
 }
 
 /**
- * Set the status of network port with id portID.
+ * @brief Update the last known status tag of a network input action.
+ *
+ * This function is similar to update_last_known_status_on_input_port, but
+ * it is called when a PTAG is granted and an upstream transient federate is not
+ * connected. It updates the last known status tag of the network input action
+ * so that it will not wait for a message or absent message from the upstream federate.
+ *
+ * This function assumes the caller holds the mutex on the top-level environment,
+ * and, if the tag actually increases, it broadcasts on `lf_port_status_changed`.
  *
  * @param env The top-level environment, whose mutex is assumed to be held.
  * @param action The action associated with the network input port.
@@ -1265,7 +1273,7 @@ static void handle_provisional_tag_advance_grant() {
                env->current_tag.time - start_time, env->current_tag.microstep, _fed.last_TAG.time - start_time,
                _fed.last_TAG.microstep);
 
-  for (int i = 0; i < _lf_zero_delay_cycle_action_table_size; i++) {
+  for (size_t i = 0; i < _lf_zero_delay_cycle_action_table_size; i++) {
     if (_lf_zero_delay_cycle_upstream_disconnected[i] == true) {
       update_last_known_status_on_action(env, _lf_zero_delay_cycle_action_table[i], PTAG);
     }
@@ -1519,7 +1527,7 @@ static void handle_upstream_connected_message(void) {
   uint16_t connected = extract_uint16(buffer);
   lf_print("********* FIXME: Upstream %d connected *********\n", connected);
   // Mark the upstream as connected.
-  for (int i = 0; i < _lf_zero_delay_cycle_action_table_size; i++) {
+  for (size_t i = 0; i < _lf_zero_delay_cycle_action_table_size; i++) {
     if (_lf_zero_delay_cycle_upstream_ids[i] == connected) {
       _lf_zero_delay_cycle_upstream_disconnected[i] = false;
     }
@@ -1538,7 +1546,7 @@ static void handle_upstream_disconnected_message(void) {
   uint16_t disconnected = extract_uint16(buffer);
   lf_print("********* FIXME: Upstream %d disconnected *********\n", disconnected);
   // Mark the upstream as disconnected.
-  for (int i = 0; i < _lf_zero_delay_cycle_action_table_size; i++) {
+  for (size_t i = 0; i < _lf_zero_delay_cycle_action_table_size; i++) {
     if (_lf_zero_delay_cycle_upstream_ids[i] == disconnected) {
       _lf_zero_delay_cycle_upstream_disconnected[i] = true;
     }
